requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.dev.in'
#
anyio==4.8.0
    # via httpx
asyncstdlib==3.13.0
    # via denonavr
attrs==25.1.0
    # via denonavr
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
defusedxml==0.7.1
    # via denonavr
denonavr==1.0.1
    # via -r .\requirements.in
distlib==0.3.9
    # via virtualenv
filelock==3.17.0
    # via virtualenv
ftfy==6.3.1
    # via denonavr
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via denonavr
identify==2.6.7
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
netifaces==0.11.0
    # via denonavr
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r .\requirements.dev.in
platformdirs==4.3.6
    # via virtualenv
pre-commit==4.1.0
    # via -r .\requirements.dev.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r .\requirements.dev.in
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via -r .\requirements.in
ruff==0.9.6
    # via -r .\requirements.dev.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via anyio
urllib3==2.3.0
    # via requests
virtualenv==20.29.2
    # via pre-commit
wcwidth==0.2.13
    # via ftfy
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
